requirements.txt
six==1.11.0
numpy=1.13.3
PyYAML=3.12
scikit-learn=0.19.1
thulac=0.1.2